requirements.txt
redis~=4.5.1
revChatGPT~=2.2.8
pyyaml-erb